dhcp6: don't require a hardware address
The systemd DHCPv6 client requires a hardware address only to
determine the IAID; NM always overrides the IAID with its own and
therefore the hwaddr is not used.

Removing such requirement allows DHCPv6 to run over PPP, which is
useful with DHCPv6-PD to get a prefix from the ISP.

To test this, I set up a server with pppoe-server, radvd in stateless
mode, and the Wide DHCPv6 server providing an address and a prefix. On
the client, NM was able to obtain a prefix using both dhcp=dhclient
and dhcp=systemd.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/478